update on confirm and handling of delete event

+++ src/js/ui/post/delete.js
@@ -4,13 +4,25 @@
 
 import { deletePost } from "../../api/post/delete";
 
+const loggedUser = JSON.parse(localStorage.getItem("adminUser"));
+
 export function onDeletePost(event) {
   event.preventDefault();
 
   const postDataId = event.target.getAttribute("data-id");
 
   if (postDataId) {
-    deletePost(postDataId);
+    const deleteThisPost = confirm("Wanna delete this post?");
+
+    if (deleteThisPost) {
+      deletePost(postDataId);
+      window.location.href = `/profile/?profile=${loggedUser.name}`;
+
+      alert("Post successfully deleted");
+    } else {
+      alert("Post deletion cancelled");
+      return;
+    }
   }
 }
